move registry to theme provider
